+++ skimage/transform/__init__.py
@@ -5,13 +5,14 @@ from .radon_transform import (radon, iradon, iradon_sart,
                               order_angles_golden_ratio)
 from .finite_radon_transform import frt2, ifrt2
 from .integral import integral_image, integrate
-from ._geometric import (warp, warp_coords, estimate_transform,
+from ._geometric import (estimate_transform,
                          matrix_transform, EuclideanTransform,
                          SimilarityTransform, AffineTransform,
                          ProjectiveTransform, FundamentalMatrixTransform,
                          EssentialMatrixTransform, PolynomialTransform,
                          PiecewiseAffineTransform)
-from ._warps import swirl, resize, rotate, rescale, downscale_local_mean
+from ._warps import (swirl, resize, rotate, rescale, downscale_local_mean,
+                     warp, warp_coords)
 from .pyramids import (pyramid_reduce, pyramid_expand,
                        pyramid_gaussian, pyramid_laplacian)
 from .seam_carving import seam_carve
